pyproject.toml
[project]
name = "janus"
version = "0.1.0"
description = "Distributed Asset Management System"
requires-python = ">=3.11"
dependencies = [
    "vnpy>=4.3.0",
    "vnpy_rpcservice",
    "vnpy_ib",
    "prompt_toolkit>=3.0.36",
    "pyyaml",
    "rich"  # 用于美化表格输出
]

[tool.setuptools.packages.find]
where = ["src"]
